// src/interfaces/ActorToMovie.tsx
// Generated by https://quicktype.io

export interface ErrorFormActorMovie {
    message: string;
    data:    Data;
}

export interface Data {
    errors: ErrorsForm;
}

export interface ErrorsForm {
    idMovie: ID;
    idActor: ID;
}

export interface ID {
    msg:      string;
    param:    string;
    location: string;
}

export interface DataActorMovie {
    idMovie:string;
    idActor:string;
}