// src/interfaces/Users.tsx
// Generated by https://quicktype.io
// Generated by https://quicktype.io

export interface UserLogin {
    msg:  string;
    data: Data;
}

export interface Data {
    user:  User;
    token: string;
}

export interface User {
    id:    number;
    name:  string;
    email: string;
}




export interface FormRegister {
    message: string;
    data:    Data;
}

export interface Data {
    errors: ErrorRegister;
}

export interface ErrorRegister {
    name:  Paras;
    email: Paras;
    pass:  Paras;
}

export interface Paras {
    msg:      string;
    param:    string;
    location: string;
}


// Generated by https://quicktype.io

export interface FormLoginError {
    msg:  string;
    data: string;
}